+++ lib/AST/OperatorNameLookup.cpp
@@ -20,6 +20,7 @@
 #include "swift/AST/DiagnosticsSema.h"
 #include "swift/AST/ImportCache.h"
 #include "swift/AST/NameLookupRequests.h"
+#include "swift/AST/SourceFile.h"
 
 #define DEBUG_TYPE "operator-name-lookup"
