+++ lib/AST/SubstitutionMap.cpp
@@ -31,6 +31,7 @@
 #include "swift/AST/Module.h"
 #include "swift/AST/ProtocolConformance.h"
 #include "swift/AST/Types.h"
+#include "swift/Basic/Defer.h"
 #include "llvm/Support/Debug.h"
 
 using namespace swift;
