+++ lib/SILOptimizer/SemanticARC/SemanticARCOptVisitor.cpp
@@ -18,6 +18,7 @@
 //===----------------------------------------------------------------------===//
 
 #include "SemanticARCOptVisitor.h"
+#include "swift/Basic/Defer.h"
 #include "swift/SIL/DebugUtils.h"
 
 using namespace swift;
