+++ lib/SILOptimizer/Utils/Generics.cpp
@@ -19,6 +19,7 @@
 #include "swift/AST/SemanticAttrs.h"
 #include "swift/AST/TypeCheckRequests.h"
 #include "swift/AST/TypeMatcher.h"
+#include "swift/Basic/Defer.h"
 #include "swift/Basic/Statistic.h"
 #include "swift/Demangling/ManglingMacros.h"
 #include "swift/SIL/DebugUtils.h"
